Fix an integer overflow

+++ src/zaldy_pmmg.c
@@ -159,7 +159,7 @@ int PMMG_memOption_memRepartition(MMG5_pMesh mesh,MMG5_pSol met) {
   npadd = (int) ( (double)avMem/bytes );
   mesh->npmax = MG_MIN(mesh->npmax,mesh->np+npadd);
   mesh->xpmax = MG_MIN(mesh->xpmax,mesh->xp+npadd);
-  mesh->nemax = MG_MIN(mesh->nemax,6*npadd+mesh->ne);
+  mesh->nemax = MG_MIN(mesh->nemax,(size_t)(6*npadd+mesh->ne));
   mesh->xtmax = MG_MIN(mesh->xtmax,ctri*npadd+mesh->xt);
 
   met->npmax  = mesh->npmax;
